+++ database/migrations/2019_07_11_193251_create_lessons_table.php
@@ -15,7 +15,7 @@ class CreateLessonsTable extends Migration
     {
         Schema::create('lessons', function (Blueprint $table) {
                 $table->bigIncrements('id');
-                $table->integer('course_id')->unsigned();
+                $table->bigInteger('course_id')->unsigned();
                 $table->string('title')->nullable();
                 $table->string('slug')->nullable();
                 $table->string('thumbnail')->nullable();
@@ -30,8 +30,8 @@ class CreateLessonsTable extends Migration
                 $table->timestamps();
                 $table->softDeletes();
 
-                $table->integer('created_by')->unsigned();
-                $table->integer('updated_by')->unsigned();
+                $table->bigInteger('created_by')->unsigned();
+                $table->bigInteger('updated_by')->unsigned();
                 
                 $table->index(['deleted_at']);
                 $table->foreign('course_id')->references('id')->on('courses')->onDelete('cascade');
